feat: improve API by returning a Zeno object from auth to issue uploads
The API has been changed so that after authentication, the user gets a Zeno object to issue API commands on. This object automatically manages user handling and token refresh.
